move `ReactHooks` to `common/`
